Add support for 3.23.0-beta.1 to match a peerDep of `^3.20.0`.
Based on https://github.com/npm/node-semver#prerelease-tags:

> If a version has a prerelease tag (for example, 1.2.3-alpha.3) then it
> will only be allowed to satisfy comparator sets if at least one
> comparator with the same [major, minor, patch] tuple also has a
> prerelease tag.

That means that if a project that uses `validatePeerDependencies`
has a peer depenency of `ember-source: ^3.0.0` and we found
`3.2.0-beta.1` `semver.satisifies` will _by default_
return `false`.

For this libraries purpose however, we need to ensure that is allowed.
Mostly because the "end user" (the host project, or the package that
included the package that uses `validatePeerDependencies`) has
explicitly asked for a prerelease version.